backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Sean Lau <lthon0225@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.75.1"
uvicorn = {extras = ["standard"], version = "^0.17.6"}
black = "^22.3.0"
pydantic = "^1.9.0"
Django = "3.2"
djangorestframework = "^3.13.1"
djangorestframework-simplejwt = "4.8.0"
djoser = "^2.1.0"
dj-rest-auth = "^2.2.4"
django-allauth = "^0.50.0"
django-cors-headers = "^3.11.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
